loading and unloading with json

+++ AP_Project/building.h
@@ -7,12 +7,12 @@ using namespace std;
 class building:public have_file
 {
 protected:
-    long double base_price;
+    double base_price;
     long long total_area;
     QString address;
     QString picture;
 public:
-    long double get_base_price(){return base_price;}
+    double get_base_price(){return base_price;}
 };
 
 #endif // BUILDING_H
